+++ frontend/src/main/resources/webapp/src/app/main/authors/cards/author-card-mini/author-card-mini.component.ts
@@ -1,0 +1,18 @@
+import {Component, Input, OnInit} from '@angular/core';
+import {Author} from "../../author";
+
+@Component({
+  selector: 'app-author-card-mini',
+  templateUrl: './author-card-mini.component.html',
+  styleUrls: ['./author-card-mini.component.scss']
+})
+export class AuthorCardMiniComponent implements OnInit {
+
+  @Input() author: Author
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}