+++ frontend/src/main/resources/webapp/src/app/main/books/books.module.ts
@@ -12,6 +12,9 @@ import {BookCardSmallComponent} from "./cards/book-card-small/book-card-small.co
 import {BookTableComponent} from './books-home/book-table/book-table.component';
 import {BookGridComponent} from './books-home/book-grid/book-grid.component';
 import {OpinionsModule} from "../opinions/opinions.module";
+import {BookCardWideComponent} from './cards/book-card-wide/book-card-wide.component';
+import {AuthorsModule} from "../authors/authors.module";
+import {AddToMyListButtonComponent} from './cards/add-to-my-list-button/add-to-my-list-button.component';
 
 @NgModule({
   imports: [
@@ -19,12 +22,14 @@ import {OpinionsModule} from "../opinions/opinions.module";
     OntimizeWebModule,
     BooksRoutingModule,
     SharedModule,
-    OpinionsModule
+    OpinionsModule,
+    AuthorsModule
   ],
     exports: [
         BookCardSmallComponent,
-        BookCardMediumComponent
-    ],
+        BookCardMediumComponent,
+        BookCardWideComponent
+  ],
     declarations: [
         BooksHomeComponent,
         BooksDetailComponent,
@@ -32,7 +37,9 @@ import {OpinionsModule} from "../opinions/opinions.module";
         BookCardSmallComponent,
         BookCardMediumComponent,
         BookTableComponent,
-        BookGridComponent
+        BookGridComponent,
+        BookCardWideComponent,
+        AddToMyListButtonComponent
     ]
 })
 export class BooksModule {