+++ frontend/src/main/resources/webapp/src/app/main/books/cards/add-to-my-list-button/add-to-my-list-button.component.spec.ts
@@ -1,0 +1,25 @@
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+
+import {AddToMyListButtonComponent} from './add-to-my-list-button.component';
+
+describe('AddToMyListButtonComponent', () => {
+  let component: AddToMyListButtonComponent;
+  let fixture: ComponentFixture<AddToMyListButtonComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ AddToMyListButtonComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AddToMyListButtonComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});