+++ frontend/src/main/resources/webapp/src/app/main/genres/genres-home/genre-grid/genre-grid.component.spec.ts
@@ -1,0 +1,25 @@
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+
+import {GenreGridComponent} from './genre-grid.component';
+
+describe('GenreGridComponent', () => {
+  let component: GenreGridComponent;
+  let fixture: ComponentFixture<GenreGridComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ GenreGridComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(GenreGridComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});