+++ frontend/src/main/resources/webapp/src/app/main/genres/genres-home/genres-home.component.ts
@@ -1,29 +1,23 @@
-import {AfterViewInit, Component, OnInit, ViewChild} from '@angular/core';
-import {OTableButtonComponent} from "ontimize-web-ngx";
-import {Router} from "@angular/router";
+import {Component, OnInit} from '@angular/core';
 
 @Component({
   selector: 'app-genre-home',
   templateUrl: './genres-home.component.html',
   styleUrls: ['./genres-home.component.scss']
 })
-export class GenresHomeComponent implements OnInit, AfterViewInit {
-
-  @ViewChild('genre_top')
-  protected genre_top: OTableButtonComponent;
-
-  ngAfterViewInit() {
-    this.genre_top.onClick.subscribe(event => {
-      this.router.navigate(['/main','genres','top'])
-    });
-  }
+export class GenresHomeComponent implements OnInit {
 
 
-  constructor(private router: Router) {
 
+  tableMode:boolean = false;
+
+  constructor() {
   }
 
   ngOnInit() {
   }
 
+  switchMode() {
+    this.tableMode = !this.tableMode;
+  }
 }