+++ CHANGELOG.md
@@ -1,7 +1,57 @@
-#### v 1.0.0-RC     	====	17/07/2020
+#### v 1.0.0-FINAL     	====	24/07/2020
+
+
+#### v 1.0.0-RC.1     	====	17/07/2020
+
 
 #### v 0.0.7-SNAPSHOT	====	10/07/2020
 
+- Front:
+    - Simple Login Guard for User Section
+	- Sort and filter books by avg_rating
+	- Revamp OTableCellRendererStarRating Component
+	- Sort and filter authors by avg_rating
+	- Sort and filter publishers by avg_rating
+	- Show publisher`s avg rating on list and detail
+	- Create Opinion module
+	- Extract Opinion display to its own component
+	- Sort reviews by rating (book detail)
+    - Publishers Home design
+    - Publishers grid-table switch
+    - Genres top_grid - table switch
+	- Author Detail: Display all reviews of author's books
+	- Publisher Detail: Display all reviews of publisher's books
+	- Refresh Book Detail on Review Submit-Mod-Remove
+	- UserModule Init
+	- User/Timeline (list of all user's reviews)
+	- User My Books List:
+		- Remove book from my list
+		- Display personal Book List
+	- Book detail:
+	    - Add and remove books to/from my list
+	- Base64 pipe for images
+	- StarImage Component fix width and height
+	- Author Card Mini (on book detail)
+	- Fix Book Card Small (vertical)
+- Back: 
+	- Publishers Rating Service
+	- Opinion Service: 
+		- vAuthorOpinionsQuery,
+		- vPublisherOpinionsQuery,
+		- vBookOpinionsQuery
+ -SQL: 
+	- Publishers Rating View
+	- SQL views optimizations
+	- My private list initialization trigger on user sign up
+	- Mod table opinion: change dates to timestamps
+	- Mod table lists: change dates to timestamps
+	- Triggers:
+		- Auto update modification date on opinions UPDATES
+		- Auto update modification date on lists UPDATES
+		- Auto update modification date on list_books CREATE, UPDATE, DELETE
+		- Auto initialize personal list on user creation
+		
+		
 #### v 0.0.6-SNAPSHOT	====	03/07/2020
 
 - Front: 
@@ -23,6 +73,7 @@
     - replace credentials with environment variables
 - SQL:
     - Default value for opinion creation date
+
 
 #### v 0.0.5-SNAPSHOT	====	26/06/2020